+++ insanic/conf/global_settings.py
@@ -124,4 +124,8 @@ GATEWAY_REGISTRATION_ENABLED = True
 KONG_HOST = "kong"
 KONG_PORT = 18000
 
+KONG_ROUTE_REGEX_PRIORITY = {"local": 10,
+                             "development": 5}
+KONG_ROUTE_REGEX_DEFAULT = 10
+
 # VAULT_APPROLE_BIND_SECRET = "pull" # value can be pull or push